Add the repository to the name of all generated containers
According to the Docker docs, any image without a domain is to be
considered an image coming from the default registry, which is not
controlled by the user. So, if the build wants to use its own registry,
just prepend the image names with the correct registry so that they can
find each other during the build. In this way, the naming convention of
docker-pkg adheres to the one used by docker itself.

Change-Id: If6ca479216c045491a05863d95bcb0fefecae00d