Add support for nightly builds
In case a command-line switch --nightly is added to the command line, a
suffix is added to the version of each image, thus triggering a global
rebuild and republish.

This can be particularly useful for images that track git projects to
HEAD (like most of the images used in CI) and that need to be rebuild
periodically.

Change-Id: I46abcd66de53d964d4441c956067dbb17f59f005