Attempt to pull images before building
docker-pkg rebuilds all images not available locally. It is the case for
a developer machine when images have been build and published from
another host. A local run thus results in new images carrying the same
tag although they are different than the published ones.

The locally build image most certainly mismatch the canonical one in the
registry since Docker builds are not reproducible.

Default to always pull from the registry.
Add --no-pull to optionally disable fetching images from the registry.

The drawback is its network bound which is a problem for slow
connections, the huge advantage is the local build will be consistent
with others.

Bug: T200720
Change-Id: I5844d86054da9f64d523e54e2fb001288fddbf1e