Fix the logic of the FSM to account for the fact we allow pulling
When we allow the build step to pull the parent image, we also want not
to build that image if not present locally, as it will be pulled
anyways.

So, fix the logic with which the initial state of the FSM is determined.

Also use constants, so that we will be able to change the logic
consistently in the future.

Change-Id: Iddfcbd246d31dc547b3f77c26e866619413d0d7f