Add a prune action
This allows to remove from the docker daemon any outdated version of an
image we're managing. This will help housekeeping on build servers where
a lot of images get built constantly, much more so if constant security
upgrades are generated.

So, add an 'action' parameter to the docker-pkg cli and methods to prune
versions of an image not defined in changelogs in a specific directory.

These actions are only local for now, and don't include registry
housekeeping (which is admittedly harder to implement safely).

Bug: T207703
Change-Id: I3aa2eada6cec0cfb2daac8a46180052a795d2c2f