Warn about lack of changelog or Dockerfile.template
I often go straight for the Dockerfile.template and forget to create a
changelog file.  In such a case, the image is not included in the scan
which is quite puzzling. Warn when either changelog or
Dockerfile.template is missing.

The conditional is now a guard with an early continue, that saves a
level of indentation in the rest of the function implementation.

Change-Id: I7a04bc8ca425d3b07ee230ee2feb5b8aed05fd8f