scan and process templates in parallel
Wikimedia CI has a hundred of images, as images are found, each triggers
a query to the Docker registry. That is network I/O bound and rather
slow when processed serially.

Move the template processing to a standalone method.
Use concurrent.futures.ThreadPoolExecutor() to parallelize the queries.
Introduce configuration setting `scan_workers` (default 8) which is
passed to ThreadPoolExecutor().

Not that scan() itself defaults to 1 to accomodate with python3.4
ThreadPoolExecutor() requiring max_workers to be set.

Change-Id: Id890f2a9d2e25dbb8614bae295bd29cf9bc31856