Add an update action
This allows to automatically update all images you select via
command-line, and also all the images that depend on those, by adding a
changelog entry to all the images that will need to be updated.

This will allow you to first run 'docker-pkg update', then submit your
change to your own VCS, and finally review and build the images that
need to be updated.

Change-Id: I8ed856ad1caae296e93100c227e2ae0438203151