Never try apt_pkg when parsing control
debian.Packages() uses debian.deb822 which defaults to try to use
apt_pkg for parsing. That causes warnings to be emitted:

    UserWarning: Parsing of Deb822 data with python3-apt's apt_pkg was
    requested but this package is not importable. Is python3-apt
    installed?

The module is not published on pypi and it seems the parsing without
apt_pkg is fast enough anyway.

Instruct debian.Packages() to not attempt to use apt_pkg.

Bug: T216836
Change-Id: I974ab43dfe926dfab91c614a3b45b240f167294b