Add typing support
This entailed changing the signature of a function, hence I'm bumping
a minor version as it's a breaking change (although AFAIK not one that
will actually break any live usage of the tool).

Also add back flake8 support, and use a more modern format for
tox.ini, and move tests outside of the module.

Change-Id: Idd3b390ed3b88fedf1cf4de7f5cfebe21df040ff