Fix UX of the argument parser
With the old configuration of the argument parser, the last argument on
the command line would always be read as the "directory" positional
argument; so the error messages would be confusing when a user forgot the
directory argument. Move the last positional argument to be part of the
subparsers to fix the user experience and obtain the correct error
message.

Bug: T253131
Change-Id: I1dfaa0f82a7302825860d9bf58277825560fa94f