Add the 'uid' template helper
Given we now allow to throw an error if non-numeric USER stanzas are
created, we add an helper that will ease the work of template writers
by allowing to link a pre-determined set of usernames to UIDs.

While not ideal, this helper allows us to cover a good portion of the
use cases (for users like "root", "www-data", etc) without adding much
complexity.

Bug: T228967
Change-Id: Icbc6af2194c8970f8ad15f855f10dae697380402