Introduce DriverInterface
A DriverInterface base class is introduced to describe the
functionality a driver should have, and use it in DI in DockerImage.

This also means that a better separation of concerns is enforced:
* image.DockerImage deals with build metadata, rendering of the dockerfile,
 managing the changelog, etc.
* driver.DriverInterface only interacts with the images by building,
removing, publishing them. It is only called from within DockerImage.
* builder.ImageFSM deals with passing the driver to DockerImage

The drivers are also separated to their own submodule.

Change-Id: Iceb1a992ca7e5c49ae22c3dbc0bde8c28797b934