Be strict on undefined variables in templates
If one config file is missing some specific settings,
docker-pkg fails with a somehow cryptic error.

To have Jinja2 to complain about the undefined setting, set the
Environment behavior for undefined to be strict instead of having them
evaluated as an empty string. The result is a meaningful exception
"X is not defined" as well as the template and line number in
the stacktrace.

Bug: T297619
Change-Id: I682556fd1ce033d2c965f0c20d58d0c89629f385