Remove seed_image
seed_image was added when the idea of managing a whole dependency
tree using a single seed image seemed reasonable.

Truth is no one wants to maintain separate image trees for different
seed images, so it makes sense to remove the setting completely and
just leave base_images which is frankly clearer.

A major version is bumped because this change is backwards incompatible.

Change-Id: I3c35c52a72e3c79e5de8dae77209e1f9921fc3f1