Add missing attrs dependency
Add attrs to the list of required packages. This dependency was
introduced in 2738fdc. The use of attrs by pytest masks the need to
explicitly declare this dependency when running the test suite.

An unrelated change to sphinx config is included in this patch to avoid
a warning that is treated as an error in Wikimedia's CI related to
a deprecation in Sphinx v5.0.

Change-Id: I204508fa9370200b676de6a150e82d20e0d9e875