Change log format to get name of image being built
When building multiples images we get the build output streamed as log.
A build emitting messages 'foo' and 'bar' would result in a log line
such as:

  2025-06-13 09:32:44 [docker-pkg-build] INFO - foo (builder.py:333)
  2025-06-13 09:32:44 [docker-pkg-build] INFO - bar (builder.py:333)

The log streaming is done using a child logger having the name of the
image. Since the logger name `%(name)s` is not included in the log
format, the information is lost.

Remove the hardcoded `[docker-pkg-build]` which makes the line slightly
longer.
Similarly remove the filename and line number, it might assist when
doing development, but is rather noisy for users.

A build of an image `php83` would thus emit:

  2025-06-13 09:32:44 INFO docker_pkg.do_build.php83: foo
  2025-06-13 09:32:44 INFO docker_pkg.do_build.php83: bar

Change-Id: I6f1a443473ae92f24651fd9879b8c156d5adb2c5